Refactor flatten to be a top level command
The journal data will need augmenting with bank transaction data in order to show the Contact's information.
