+++ tests/test_main.py
@@ -15,7 +15,7 @@ from xero.auth import OAuth2PKCECredentials
 from xerotrust import export
 from xerotrust import main
 from xerotrust.authentication import authenticate, SCOPES, credentials_from_file
-from xerotrust.export import ALL_JOURNAL_KEYS
+from xerotrust.flatten import ALL_JOURNAL_KEYS
 from xerotrust.main import cli
 from xerotrust.transform import show
 from .helpers import FileChecker
@@ -1375,7 +1375,7 @@ class TestJournalsCheck:
         )
 
 
-class TestJournalsFlatten:
+class TestFlatten:
     def write_journal_file(self, path: Path, journals: list[dict[str, Any]]) -> None:
         """Helper to write a JSON Lines file."""
         path.write_text('\n'.join(json.dumps(j) for j in journals) + '\n')
@@ -1409,7 +1409,7 @@ class TestJournalsFlatten:
         ]
         self.write_journal_file(journal_file, journals_data)
 
-        result = run_cli(tmp_path, 'journals', 'flatten', str(journal_file))
+        result = run_cli(tmp_path, 'flatten', str(journal_file))
 
         # We don't use check_output here so we can see what the raw csv looks like,
         header = ','.join(ALL_JOURNAL_KEYS)
@@ -1441,7 +1441,7 @@ class TestJournalsFlatten:
         self.write_journal_file(file1, journals_data1)
         self.write_journal_file(file2, journals_data2)
 
-        result = run_cli(tmp_path, 'journals', 'flatten', str(file1), str(file2))
+        result = run_cli(tmp_path, 'flatten', str(file1), str(file2))
 
         expected_rows = [
             {'JournalID': 'j1', 'JournalNumber': 1, 'JournalLineID': 'jl1a'},
@@ -1453,7 +1453,7 @@ class TestJournalsFlatten:
         empty_file = tmp_path / "empty.jsonl"
         empty_file.touch()
 
-        result = run_cli(tmp_path, 'journals', 'flatten', str(empty_file))
+        result = run_cli(tmp_path, 'flatten', str(empty_file))
 
         self.check_output(result.output, expected=[])
 
@@ -1465,7 +1465,7 @@ class TestJournalsFlatten:
         ]
         self.write_journal_file(journal_file, journals_data)
 
-        result = run_cli(tmp_path, 'journals', 'flatten', str(journal_file))
+        result = run_cli(tmp_path, 'flatten', str(journal_file))
 
         self.check_output(result.output, expected=[])
 
@@ -1491,7 +1491,7 @@ class TestJournalsFlatten:
         ]
         self.write_journal_file(journal_file, journals_data)
 
-        result = run_cli(tmp_path, 'journals', 'flatten', str(journal_file))
+        result = run_cli(tmp_path, 'flatten', str(journal_file))
 
         expected_rows = [
             {
@@ -1523,7 +1523,6 @@ class TestJournalsFlatten:
 
         run_cli(
             tmp_path,
-            'journals',
             'flatten',
             str(journal_file),
             '--output',
@@ -1555,7 +1554,7 @@ class TestJournalsFlatten:
         ]
         self.write_journal_file(journal_file, journals_data)
 
-        result = run_cli(tmp_path, 'journals', 'flatten', str(journal_file))
+        result = run_cli(tmp_path, 'flatten', str(journal_file))
 
         expected_rows = [
             {
